services/tester/requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-dotenv==1.0.1
python-owasp-zap-v2.4==0.1.0
requests==2.32.3
six==1.16.0
urllib3==2.2.2
Werkzeug==3.0.4
zaproxy==0.3.2
